vercel build failure refactor
