+++ src/App.tsx
@@ -1,7 +1,7 @@
 
 import React, { useEffect, useState } from 'react';
 import './App.css'
-import HeaderComponent from './components/Header'
+import HeaderComponent from './components/HeaderComponent'
 import LeftPanelComponent from './components/LeftPanelComponent';
 import { fetchGenres, fetchAllTitle } from './services/ProductService';
 import TitlePanelComponent from './components/TitlePannelComponent';
